+++ koku/api/report/test/ocp_aws/tests_views.py
@@ -15,6 +15,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 """Test the OCP on AWS Report views."""
+import logging
 import datetime
 import random
 from urllib.parse import quote_plus, urlencode
@@ -33,6 +34,8 @@ from api.query_handler import TruncDayString
 from api.report.test.ocp_aws.helpers import OCPAWSReportDataGenerator
 from api.utils import DateHelper
 from reporting.models import OCPAWSCostLineItemDailySummary
+
+LOG = logging.getLogger(__name__)
 
 
 class OCPAWSReportViewTest(IamTestCase):
@@ -337,6 +340,8 @@ class OCPAWSReportViewTest(IamTestCase):
 
         url = url + '?' + urlencode(params, quote_via=quote_plus)
         response = client.get(url, **self.headers)
+        LOG.critical('XXX1: %s', url)
+        LOG.critical('XXX2: %s', response)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
         data = response.json()
